docs(tokens): ensure internal links from tokens page don't remove the url path (e.g. /main/)
